refactored recList.js and converted scripts to type=module
