+++ public/js/recList.js
@@ -1,44 +1,13 @@
+import {
+  createErrorList,
+  deleteError,
+  addClassToArr,
+  removeClassFromArr,
+} from "./helpers.js";
+
 /**
  * HELPERS
  */
-
-const createErrorList = (errors) => {
-  const ul = document.createElement("ul");
-  ul.classList.add("clienterror");
-  for (const error of errors) {
-    const errorLi = document.createElement("li");
-    errorLi.classList.add("no-bullet");
-    const errorP = document.createElement("p");
-    errorP.textContent = error;
-    errorLi.appendChild(errorP);
-    ul.appendChild(errorLi);
-  }
-  return ul;
-};
-
-const deleteError = () => {
-  const errorLi = document.querySelector(".clienterror");
-  if (errorLi) errorLi.remove();
-  // Also check for server generated errors
-  const serverErrorEl = document.querySelector(".serverErrorContainer");
-  if (serverErrorEl) serverErrorEl.remove();
-};
-
-const addClassTo = (el, className) => {
-  el.classList.add(className);
-};
-
-const removeClassFrom = (el, className) => {
-  el.classList.remove(className);
-};
-
-const addClassToArr = (elArr, className) => {
-  elArr.forEach((el) => el.classList.add(className));
-};
-
-const removeClassFromArr = (elArr, className) => {
-  elArr.forEach((el) => el.classList.remove(className));
-};
 
 const fillStars = (stars, maxStarValue) => {
   const filledStars = stars.filter((star) => star.value <= maxStarValue);
@@ -49,6 +18,20 @@ const fillStars = (stars, maxStarValue) => {
   addClassToArr(unfilledStars, "fa-regular");
 };
 
+const ajaxPost = async (url, parentEl) => {
+  return await $.ajax({
+    method: "POST",
+    url,
+  })
+    .then(() => {
+      location.reload();
+    })
+    .fail((xhr, _, err) => {
+      const errEl = createErrorList([`Status ${xhr.status}: ${err}`]);
+      parentEl.appendChild(errEl);
+    });
+};
+
 /**
  * DOM ELEMENTS
  */
@@ -56,37 +39,18 @@ const fillStars = (stars, maxStarValue) => {
 const main = document.querySelector("main");
 const addFriendForm = document.querySelector("#addFriendForm");
 const addReviewForm = document.querySelector("#addReviewForm");
-
-const star1 = document.querySelector("#star1");
-const star2 = document.querySelector("#star2");
-const star3 = document.querySelector("#star3");
-const star4 = document.querySelector("#star4");
-const star5 = document.querySelector("#star5");
-
-const stars = [star1, star2, star3, star4, star5];
-
-/**
- * CONSTANTS
- */
+const stars = [...document.querySelectorAll(".fa-star")];
 
 /**
  * RUNTIME Dynamic HTML & CSS
  */
 
-const filledStars = stars.filter(
-  (star) => star.id[star.id.length - 1] <= handlebars.REVIEW_RATING
-);
-
-filledStars.forEach((star) => {
-  star.classList.remove("fa-regular");
-  star.classList.add("fa-solid");
-});
+fillStars(stars, handlebars.REVIEW_RATING);
 
 /**
  * EVENT LISTENERS
  */
-if (document.getElementById("addFriendForm")){
-  console.log("friend form handler being added")
+if (document.getElementById("addFriendForm")) {
   addFriendForm.addEventListener("submit", (e) => {
     e.preventDefault();
     deleteError();
@@ -94,21 +58,10 @@ if (document.getElementById("addFriendForm")){
       console.error("Author not found... Invalid page...");
       return (window.location.href = "/main");
     }
-    $.ajax({
-      method: "POST",
-      url: `/accounts/friend/${handlebars.AUTHOR_NAME}`,
-    })
-      .then(() => {
-        return (window.location.href = `/recommendations/${handlebars.REC_ID}`);
-      })
-      .fail((xhr, _, err) => {
-        const errEl = createErrorList([`Status ${xhr.status}: ${err}`]);
-        main.appendChild(errEl);
-      });
+    ajaxPost(`/accounts/friend/${handlebars.AUTHOR_NAME}`, main);
   });
 }
 if (document.getElementById("addReviewForm")) {
-  console.log("review form handler being added")
   addReviewForm.addEventListener("submit", (e) => {
     e.preventDefault();
     deleteError();
@@ -116,20 +69,11 @@ if (document.getElementById("addReviewForm")) {
       console.error("Recommendation ID not found... Invalid page...");
       return (window.location.href = "/main");
     }
-
     const rating = e.submitter.value;
-
-    $.ajax({
-      method: "POST",
-      url: `/recommendations/review/${handlebars.REC_ID}?rating=${rating}`,
-    })
-      .then((res) => {
-        return (window.location.href = `/recommendations/${handlebars.REC_ID}`);
-      })
-      .fail((xhr, _, err) => {
-        const errEl = createErrorList([`Status ${xhr.status}: ${err}`]);
-        main.appendChild(errEl);
-      });
+    ajaxPost(
+      `/recommendations/review/${handlebars.REC_ID}?rating=${rating}`,
+      main
+    );
   });
 
   addReviewForm.addEventListener("mouseover", (e) => {
