+++ views/accounts.handlebars
@@ -168,4 +168,4 @@
   </div>
 </main>
 
-<script defer src="/public/js/accounts.js"></script>
+<script type="module" src="/public/js/accounts.js"></script>
