+++ views/friends.handlebars
@@ -151,4 +151,4 @@
   </div>
 </main>
 
-<script defer src="/public/js/friends.js"></script>
+<script type="module" src="/public/js/friends.js"></script>
