+++ views/layouts/main.handlebars
@@ -41,7 +41,6 @@ don't forget to include your css and your client-side script file references in 
     ></script>
 
     <link rel="stylesheet" href="/public/css/styles.css" />
-    <script defer src="/public/js/navbar.js"></script>
   </head>
   <body>
     <nav
@@ -88,8 +87,6 @@ don't forget to include your css and your client-side script file references in 
         {{{body}}}
       </div>
     </div>
-    <script
-      src="https://rawgit.com/leizongmin/js-xss/master/dist/xss.js"
-    ></script>
+    <script type="module" src="/public/js/navbar.js"></script>
   </body>
 </html>
