+++ views/login.handlebars
@@ -40,4 +40,4 @@
   {{/if}}
 </main>
 
-<script defer src="/public/js/auth.js"></script>
+<script type="module" src="/public/js/auth.js"></script>
