+++ views/recommendationList.handlebars
@@ -21,9 +21,6 @@
             class="fa-solid fa-user-group rounded-pill border-0 p-3"
           >
           </button>
-          {{! <script>
-            console.log("bad");
-          </script> }}
         </form>
       {{/if}}
     </div>
@@ -147,4 +144,4 @@
   REC_ID: '{{recId}}', REVIEW_RATING: '{{reviewRating}}', FRIEND_FORM: '{{showFriendButton}}'
   };
 </script>
-<script src="/public/js/recList.js"></script>
+<script src="/public/js/recList.js" type="module"></script>
